+++ webapp/src/main/java/com/uniovi/wichatwebapp/entities/Game.java
@@ -1,17 +1,11 @@
 package com.uniovi.wichatwebapp.entities;
 
+import com.uniovi.wichatwebapp.services.QuestionService;
 import entities.Question;
 import entities.QuestionCategory;
 
-public class Game {
-    private int maxNumberOfQuestions = 10;
-    private int points;
-    private int questions = 0;
-    private int rightAnswers = 0;
-    private int wrongAnswers = 0;
-    private Question currentQuestion;
-    private QuestionCategory category;
-    private int timer = 30;
+public class Game extends AbstractGame{
+    private final QuestionCategory category;
 
     public Game(QuestionCategory category) {
         this.category = category;
@@ -19,63 +13,18 @@ public class Game {
 
     public Game(QuestionCategory category, int timer, int maxNumberOfQuestions) {
         this.category = category;
-        this.timer = timer;
-        this.maxNumberOfQuestions = maxNumberOfQuestions;
-    }
-
-    public boolean checkAnswer(String id) {
-        return currentQuestion.getCorrectAnswer().getId().equals(id);
-    }
-
-    public boolean hasGameEnded() {
-        return questions >= maxNumberOfQuestions;
-    }
-
-    public void correctAnswer(){
-        rightAnswers++;
-        points+=100;
-        questions++;
-    }
-
-    public void wrongAnswer(){
-        wrongAnswers++;
-        points-=25;
-        questions++;
-    }
-
-    public int getPoints() {
-        return points;
-    }
-
-    public int getQuestions() {
-        return questions;
-    }
-
-    public int getRightAnswers() {
-        return rightAnswers;
-    }
-
-    public int getWrongAnswers() {
-        return wrongAnswers;
-    }
-
-    public Question getCurrentQuestion() {
-        return currentQuestion;
-    }
-
-    public int getMaxNumberOfQuestions() {
-        return maxNumberOfQuestions;
-    }
-
-    public void setCurrentQuestion(Question currentQuestion) {
-        this.currentQuestion = currentQuestion;
-    }
-
-    public int getTimer() {
-        return timer;
+        setTimer(timer);
+        setMaxNumberOfQuestions(maxNumberOfQuestions);
     }
 
     public QuestionCategory getCategory() {
         return category;
+    }
+
+    @Override
+    public void nextQuestion(QuestionService questionService){
+        Question question = questionService.getRandomQuestion(this.category);
+        setCurrentQuestion(question);
+        questionService.removeQuestion(question);
     }
 }
