+++ webapp/src/main/java/com/uniovi/wichatwebapp/entities/GameMultipleCategories.java
@@ -1,0 +1,15 @@
+package com.uniovi.wichatwebapp.entities;
+
+import com.uniovi.wichatwebapp.services.QuestionService;
+import entities.Question;
+
+public class GameMultipleCategories extends AbstractGame{
+
+
+    @Override
+    public void nextQuestion(QuestionService questionService) {
+        Question question = questionService.getRandomQuestionNoCategory();
+        setCurrentQuestion(question);
+        questionService.removeQuestion(question);
+    }
+}
