+++ webapp/src/main/resources/templates/home.html
@@ -22,7 +22,9 @@
     <a href="/game/start/BIOLOGY" class="game-link big-button-anchor" sec:authorize="isAuthenticated()">
         Biology Game
     </a>
-
+    <a href="/game/start/CATEGORIES" class="game-link big-button-anchor" sec:authorize="isAuthenticated()">
+        Multiple Categories Game
+    </a>
     <a href="/game/personalized" class="game-link big-button-anchor" sec:authorize="isAuthenticated()">
         Custom Game
     </a>
